move context to context.py, add command.py
